requirements.txt
alembic==0.9.1
aniso8601==1.2.1
appdirs==1.4.3
click==6.7
Flask==0.12.1
Flask-HTTPAuth==3.2.2
Flask-JWT==0.3.2
Flask-Login==0.4.0
Flask-Migrate==2.0.3
Flask-RESTful==0.3.5
flask-restplus==0.10.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
functools32==3.2.3.post2
itsdangerous==0.24
Jinja2==2.9.6
JsonForm==0.0.2
jsonschema==2.6.0
JsonSir==0.0.2
Mako==1.0.6
MarkupSafe==1.0
nose==1.3.7
packaging==16.8
psycopg2==2.7.1
PyJWT==1.4.2
pyparsing==2.2.0
python-dateutil==2.6.0
Python-EasyConfig==0.1.6
python-editor==1.0.3
pytz==2017.2
PyYAML==3.12
Resource==0.2.0
six==1.9.0
SQLAlchemy==1.1.9
Werkzeug==0.12.1
